pyproject.toml
[project]
name = "itmo-chatbot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "crawl4ai>=0.7.2",
    "pdfplumber>=0.11.7",
    "playwright>=1.54.0",
    "pydantic>=2.11.7",
    "pydantic-ai>=0.4.11",
    "pydantic-settings>=2.10.1",
    "python-telegram-bot>=22.3",
    "typer>=0.16.0",
]
